+++ src/timeseries/spectrum.jl
@@ -32,10 +32,10 @@ function pspectrum(x::AbstractDimVector, spec::Spectrogram; kwargs...)
 end
 
 function pspectrum(x::AbstractDimArray, spec::Spectrogram; query=Ti, kwargs...)
-    times = SPEDAS.times(x)
+    ts = times(x)
     dims = otherdims(x, query)
     specs = map(eachslice(x; dims)) do slice
-        pspectrum(slice, times, spec; kwargs...)
+        pspectrum(slice, ts, spec; kwargs...)
     end
     cat(specs...; dims)
 end